Fix create restore editors not being recyclable
