+++ ui/page/components/components.go
@@ -11,6 +11,7 @@ import (
 	"strings"
 	"time"
 
+	"gioui.org/io/key"
 	"gioui.org/layout"
 	"gioui.org/unit"
 
@@ -839,4 +840,29 @@ func CreateOrUpdateWalletDropDown(l *load.Load, dwn **decredmaterial.DropDown, w
 func CreateOrderDropDown(l *load.Load) *decredmaterial.DropDown {
 	return l.Theme.DropDown([]decredmaterial.DropDownItem{{Text: values.String(values.StrNewest)},
 		{Text: values.String(values.StrOldest)}}, 1)
+}
+
+//Tab key event handler for pages withe ditors
+func handleTabEvent(event chan *key.Event) bool {
+	var isTabPressed bool
+	select {
+	case event := <-event:
+		if event.Name == key.NameTab && event.State == key.Press {
+			isTabPressed = true
+		}
+	default:
+	}
+	return isTabPressed
+}
+
+func HandleTabEvent(event chan *key.Event) bool {
+	var isTabPressed bool
+	select {
+	case event := <-event:
+		if event.Name == key.NameTab && event.State == key.Press {
+			isTabPressed = true
+		}
+	default:
+	}
+	return isTabPressed
 }
