+++ ui/page/utils.go
@@ -16,6 +16,7 @@ import (
 	"gioui.org/widget"
 	"github.com/planetdecred/dcrlibwallet"
 	"github.com/planetdecred/godcr/ui/decredmaterial"
+	"github.com/planetdecred/godcr/ui/page/components"
 	"github.com/planetdecred/godcr/ui/values"
 	"github.com/planetdecred/godcr/wallet"
 )
@@ -121,15 +122,28 @@ func handleSubmitEvent(editors ...*widget.Editor) bool {
 	return submit
 }
 
-//Tab key event handler for pages withe ditors
-func handleTabEvent(event chan *key.Event) bool {
-	var isTabPressed bool
-	select {
-	case event := <-event:
-		if event.Name == key.NameTab && event.State == key.Press {
-			isTabPressed = true
+func SwitchEditors(keyEvent chan *key.Event, editors ...*widget.Editor) {
+	for i := 0; i < len(editors); i++ {
+		if editors[i].Focused() {
+			if components.HandleTabEvent(keyEvent) {
+				if i == len(editors)-1 {
+					editors[0].Focus()
+				} else {
+					editors[i+1].Focus()
+				}
+			}
 		}
-	default:
 	}
-	return isTabPressed
+}
+
+func SwitchRestoreEditors(editors ...*widget.Editor) {
+	for i := 0; i < len(editors); i++ {
+		if editors[i].Focused() {
+			if i == len(editors)-1 {
+				editors[0].Focus()
+			} else {
+				editors[i+1].Focus()
+			}
+		}
+	}
 }
